+++ packages/core/src/languages/html/css-values.ts
@@ -54,10 +54,11 @@ function newStyleDeclarationValueExpressions(
  * @param options The {@link CssDeclarationValueOptions}.
  * @returns A set of {@link MangleExpression}s.
  * @since v0.1.14
+ * @version v0.1.17
  */
 export default function cssDeclarationValueExpressionFactory(
   options: CssDeclarationValueOptions,
-): MangleExpression[] {
+): Iterable<MangleExpression> {
   const valuePrefix = options.prefix ? options.prefix : "";
   const valueSuffix = options.suffix ? options.suffix : "";