+++ packages/core/src/languages/utils/simple-language-plugin.class.ts
@@ -5,8 +5,9 @@ import type { MangleExpression, WebManglerLanguagePlugin } from "../../types";
  * options.
  *
  * @since v0.1.14
+ * @version v0.1.17
  */
-export type ExpressionFactory = (options: any) => MangleExpression[]; // eslint-disable-line @typescript-eslint/no-explicit-any
+export type ExpressionFactory = (options: any) => Iterable<MangleExpression>; // eslint-disable-line @typescript-eslint/no-explicit-any
 
 /**
  * The {@link SimpleLanguagePlugin} abstract class provides an implementation of
@@ -26,7 +27,7 @@ export default abstract class SimpleLanguagePlugin
    * The languages, including aliases, that this {@link SimpleLanguagePlugin}
    * supports.
    */
-  private readonly languages: string[];
+  private readonly languages: Iterable<string>;
 
   /**
    * A map from {@link MangleExpression}-set names to a functions that produce
@@ -47,10 +48,10 @@ export default abstract class SimpleLanguagePlugin
    * @param languages Supported language, including aliases.
    * @param expressionFactories The {@link ExpressionFactory}s to use.
    * @since v0.1.15
-   * @version v0.1.15
+   * @version v0.1.17
    */
   constructor(
-    languages: string[],
+    languages: Iterable<string>,
     expressionFactories: Map<string, ExpressionFactory>,
   ) {
     this.languages = languages;
@@ -59,12 +60,13 @@ export default abstract class SimpleLanguagePlugin
 
   /**
    * @inheritDoc
+   * @version v0.1.17
    */
   getExpressions(
     name: string,
     options: unknown,
-  ): Map<string, MangleExpression[]> {
-    const map: Map<string, MangleExpression[]> = new Map();
+  ): Map<string, Iterable<MangleExpression>> {
+    const map: Map<string, Iterable<MangleExpression>> = new Map();
 
     const expressionFactory = this.expressionFactories.get(name);
     if (expressionFactory === undefined) {
@@ -72,7 +74,11 @@ export default abstract class SimpleLanguagePlugin
     }
 
     const expressions = expressionFactory(options);
-    return this.languages.reduce((m, lang) => m.set(lang, expressions), map);
+    for (const language of this.languages) {
+      map.set(language, expressions);
+    }
+
+    return map;
   }
 
   /**
@@ -80,8 +86,9 @@ export default abstract class SimpleLanguagePlugin
    * SimpleLanguagePlugin} was initialized.
    *
    * @inheritDoc
+   * @version v0.1.17
    */
-  getLanguages(): string[] {
+  getLanguages(): Iterable<string> {
     return this.languages;
   }
 }