+++ packages/core/CHANGELOG.md
@@ -110,5 +110,5 @@ Versioning].
 
 - Initial release with 4 built-in manglers and 3 built-in supported languages.
 
-[keep a changelog]: https://keepachangelog.com/en/1.0.0/
-[semantic versioning]: https://semver.org/spec/v2.0.0.html
+[keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
+[semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"