+++ packages/core/CHANGELOG.md
@@ -8,7 +8,9 @@ Versioning].
 
 ## [Unreleased]
 
+- Generalize interfaces to allow for other iterable types than arrays.
 - Support mangling attributes in `attr()` in HTML style attribute values.
+- All inputted files will now always be returned, even if not mangled.
 
 ## [0.1.16] - 2021-03-19